fix(#1577): release-active token-order guard on single-gen lazy LIMIT prefix
The trusted full-cap fast path in capped_fallback_scan returns scan_stream's
first `cap` rows directly. Its only correctness check was a debug_assert, so a
RELEASE build ran with NO guard and could silently return WRONG rows if the
single-generation lazy scan_stream pipeline ever diverged from the authoritative
token-sorted scan (multi-generation is prefix-correct by construction).

- Add prefix_is_token_ordered: an O(cap) (token, key)-monotonicity check that
  runs on EVERY build. It never decodes the tail, so D1's decode-stop is intact.
- On a detected divergence: log loudly and fall back to the authoritative scan
  via the shared reconcile_via_authoritative_scan helper (never return the
  possibly-wrong prefix). Keep the exhaustive debug_assert for the residual case
  the O(cap) check can't see (a transposition straddling the cap).
- Split limit_pushdown.rs -> limit_pushdown/{mod,tests}.rs (campsite rule #1116).
- TDD: white-box guard unit tests (fail-first, revert-verified) + a single-gen
  end-to-end parity test that runs in a normal (release-representative) build.
